Created training boss
Made TrainingDummy, a boss for training maps that does not move and just hurls insults at players.  He has 1 HP, and his sole purpose is to be annoying enough to motivate the players to move to him and smack him.  Also created the effect he puts on players, `Embarrassed` in debuffs.